Updates to steps 4 & 5
