+++ tutorials/nservicebus-step-by-step/4-publishing-events/Solution/ClientUI/InputLoopService.cs
@@ -8,7 +8,7 @@ using Microsoft.Extensions.Hosting;
 
 namespace ClientUI;
 
-public class Worker(IMessageSession messageSession, ILogger<Worker> logger) : BackgroundService
+public class InputLoopService(IMessageSession messageSession, ILogger<InputLoopService> logger) : BackgroundService
 {
     protected override async Task ExecuteAsync(CancellationToken stoppingToken)
     {
