+++ tutorials/nservicebus-step-by-step/4-publishing-events/Solution/ClientUI/Program.cs
@@ -22,7 +22,7 @@ class Program
 
         builder.UseNServiceBus(endpointConfiguration);
 
-        builder.Services.AddHostedService<Worker>();
+        builder.Services.AddHostedService<InputLoopService>();
 
         await builder.Build().RunAsync();
     }
