--- tutorials/nservicebus-step-by-step/5-retrying-errors/Snippets/Core_8/OrderPlaced.cs
@@ -1,10 +1,0 @@
-﻿using NServiceBus;
-
-namespace Core_8
-{
-    public class OrderPlaced :
-        IEvent
-    {
-        public string OrderId { get; set; }
-    }
-}
