--- tutorials/nservicebus-step-by-step/5-retrying-errors/Snippets/Core_8/PlaceOrder.cs
@@ -1,10 +1,0 @@
-﻿using NServiceBus;
-
-namespace Core_8
-{
-    public class PlaceOrder :
-        ICommand
-    {
-        public string OrderId { get; set; }
-    }
-}
