+++ tutorials/nservicebus-step-by-step/5-retrying-errors/Snippets/Core_9/OrderPlaced.cs
@@ -1,0 +1,9 @@
+﻿using NServiceBus;
+
+namespace Core_9;
+
+public class OrderPlaced :
+    IEvent
+{
+    public string OrderId { get; set; }
+}
