+++ tutorials/nservicebus-step-by-step/5-retrying-errors/Snippets/Snippets.sln
@@ -1,19 +1,24 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29728.190
+# Visual Studio Version 17
+VisualStudioVersion = 17.11.35303.130
 MinimumVisualStudioVersion = 15.0.26730.12
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core_8", "Core_8\Core_8.csproj", "{10CE989A-8E4F-400C-8BBD-3935B9FEE695}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core_9", "Core_9\Core_9.csproj", "{BB935DF8-5428-4C4E-BFBE-540CDA8F50C9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{10CE989A-8E4F-400C-8BBD-3935B9FEE695}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{10CE989A-8E4F-400C-8BBD-3935B9FEE695}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BB935DF8-5428-4C4E-BFBE-540CDA8F50C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BB935DF8-5428-4C4E-BFBE-540CDA8F50C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BB935DF8-5428-4C4E-BFBE-540CDA8F50C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BB935DF8-5428-4C4E-BFBE-540CDA8F50C9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8A4C7099-397A-4671-9F37-7346F3FBDC87}
 	EndGlobalSection
 EndGlobal
