+++ tutorials/nservicebus-step-by-step/5-retrying-errors/Solution/Sales/PlaceOrderHandler.cs
@@ -18,18 +18,18 @@ public class PlaceOrderHandler(ILogger<PlaceOrderHandler> logger) :
         // This is normally where some business logic would occur
 
         // Uncomment to test throwing a systemic exception
-        //throw new Exception("BOOM");
+        throw new Exception("BOOM");
 
-        // Uncomment to test throwing a transient exception
+        //Uncomment to test throwing a transient exception
         //if (random.Next(0, 5) == 0)
         //{
         //    throw new Exception("Oops");
         //}
 
-        var orderPlaced = new OrderPlaced
-        {
-            OrderId = message.OrderId
-        };
-        return context.Publish(orderPlaced);
+        //var orderPlaced = new OrderPlaced
+        //{
+        //    OrderId = message.OrderId
+        //};
+        //return context.Publish(orderPlaced);
     }
 }
