+++ tutorials/nservicebus-step-by-step/5-retrying-errors/tutorial.md
@@ -107,19 +107,21 @@ Next, run the solution.
 When we do these steps, we'll see a wall of exception messages in white text, which is log level INFO, followed by one in yellow text, which is log level WARN. The exception traces in white are the failures during immediate retries, and the last trace in yellow is the failure that hands the message over to delayed retries.
 
 ```
-INFO  Sales.PlaceOrderHandler Received PlaceOrder, OrderId = e927667c-b949-47ee-8ea2-f29523909784
-WARN  NServiceBus.RecoverabilityExecutor Delayed Retry will reschedule message '53ac6836-48ef-49dd-aabb-a67c0104a2a5' after a delay of 00:00:10 because of an exception:
-System.Exception: BOOM
-   at <stack trace>
+info: Sales.PlaceOrderHandler[0]
+      Received PlaceOrder, OrderId = a905a24c-a630-475b-af8b-452db7c95d3a
+warn: NServiceBus.DelayedRetry[0]
+      Delayed Retry will reschedule message '605705bd-3e31-4241-b56a-b20500a6c475' after a delay of 00:00:10 because of an exception:
+      System.Exception: BOOM
 ```
 
 Ten seconds later, the retries begin again, followed by another yellow trace, sending the message back to delayed retries. Twenty seconds after that, another set of traces. Finally, 30 seconds after that, the final exception trace will be shown in red, which is log level ERROR. This is where NServiceBus gives up on the message and redirects it to the error queue.
 
 ```
-INFO  Sales.PlaceOrderHandler Received PlaceOrder, OrderId = e927667c-b949-47ee-8ea2-f29523909784
-ERROR NServiceBus.RecoverabilityExecutor Moving message '53ac6836-48ef-49dd-aabb-a67c0104a2a5' to the error queue 'error' because processing failed due to an exception:
-System.Exception: BOOM
-   at < stack trace>
+info: Sales.PlaceOrderHandler[0]
+      Received PlaceOrder, OrderId = a905a24c-a630-475b-af8b-452db7c95d3a
+fail: NServiceBus.MoveToError[0]
+      Moving message '605705bd-3e31-4241-b56a-b20500a6c475' to the error queue 'error' because processing failed due to an exception:
+      System.Exception: BOOM
 ```
